Fix strict-mode GCC and UBSan failures on Linux
Development happens on macOS, so the Linux warning set had never actually
been run against this code. Building with GCC and -Werror surfaced fourteen
errors and UBSan surfaced one real bug.

Most of the noise comes from a genuine type difference: size_t is uint64_t
and int64_t is long on 64-bit Linux, while both pairs are distinct types on
macOS. The same static_cast is therefore *required* by Apple clang's
-Wconversion and *rejected* by GCC's -Wuseless-cast. Narrow<To, From>() in
types.h routes the conversion through a template so the types stay dependent,
which satisfies both while keeping the narrowing explicit and greppable.

Similarly, EAGAIN and EWOULDBLOCK are the same value on Linux, so the portable
`err == EAGAIN || err == EWOULDBLOCK` idiom reads as a tautology to
-Wlogical-op. IsWouldBlock() keeps the portability without the warning.

The UBSan finding is not cosmetic: Buffer::Append passed a null pointer to
memcpy, which is undefined even when the length is zero, and an empty ByteSpan
routinely carries a null data(). macOS happened not to reach the same call
sites.

+++ src/metadata/cluster_metadata.cc
@@ -139,7 +139,9 @@ std::vector<PartitionAssignment> ClusterMetadata::ComputeAssignments(
       std::min<std::size_t>(static_cast<std::size_t>(config.replication_factor), broker_count);
   // Offsetting by a hash of the topic name stops every topic from putting its
   // partition-0 leader on the same broker.
-  const std::uint32_t topic_offset = Crc32c(AsBytes(config.name)) % broker_count;
+  // size_t throughout: the offset is only ever used as an index into the
+  // broker list, and computing it in uint32_t narrows the modulus result.
+  const std::size_t topic_offset = Crc32c(AsBytes(config.name)) % broker_count;
 
   for (std::int32_t p = 0; p < config.partition_count; ++p) {
     PartitionAssignment assignment;
